+++ src/data/preprocessors/matchers/features/core.py
@@ -1,6 +1,6 @@
-from src.data.preprocessors.matchers.feature.brute_force import brute_force_matching_with_orb, \
+from src.data.preprocessors.matchers.features.brute_force import brute_force_matching_with_orb, \
     brute_force_matching_with_sift
-from src.data.preprocessors.matchers.feature.flann import flann_matcher
+from src.data.preprocessors.matchers.features.flann import flann_matcher
 
 
 def process_feature_mapping_approaches(template_image, input_image, image_name="1.jpg", template_name="template.jpg"):
