+++ src/data/preprocessors/matchers/features/flann.py
@@ -3,7 +3,7 @@ import os
 import cv2
 
 from src import DEBUG_FOLDER
-from src.data.preprocessors.matchers.feature.utils import make_distance_ratio_distribution, \
+from src.data.preprocessors.matchers.features.utils import make_distance_ratio_distribution, \
     extract_good_matches, apply_feature_detector, extract_good_matches_mask, detect_homography_polygon
